backend/requirements.txt
# Core FastAPI stack
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1
sse-starlette==1.6.5

# Pydantic + typing
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.14.1

# ASGI dependencies
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1
click==8.2.1
colorama==0.4.6
idna==3.10

# Environment variables
python-dotenv==1.1.1

# Database
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Tokenization
tiktoken==0.7.0

# LLM providers
openai==1.50.2
anthropic==0.58.2

# Retry logic
tenacity==8.2.3

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# PDF processing
pymupdf==1.26.6

# YouTube support
google-api-python-client==2.149.0

# Optional: Excel/CSV support
pandas==2.2.0
openpyxl==3.1.2

# Optional: Word documents
python-docx==1.1.0

# Optional: OCR support (if needed)
pillow==10.2.0
pytesseract==0.3.10

# Encryption for API keys
cryptography==41.0.7
